requirements.txt
python-speech-features
torch
numpy
matplotlib
scipy
lws
docopt
nnmnkwii
tensorboardX
pyyaml
tqdm